+++ Source/Engine/Graphics/Objects/Texture2D.cpp
@@ -5,9 +5,9 @@
 
 #include "Engine/ImageLoader.hpp"
 
-Texture2D::Texture2D(Texture2DTarget target, 
-                     const fs::path& absolute)
-  : id{ 0 }
+Texture2D::Texture2D(Texture2DTarget target,
+                     const fs::path &absolute)
+    : id{0}
 {
   Create(target);
   LoadImageData(absolute);
@@ -17,9 +17,9 @@ void Texture2D::Create(Texture2DTarget target)
 {
   glCreateTextures(static_cast<u32>(target), 1, &id);
 }
-void Texture2D::Delete() 
-{ 
-  glDeleteTextures(1, &id); 
+void Texture2D::Delete()
+{
+  glDeleteTextures(1, &id);
   id = 0;
 }
 bool Texture2D::IsValid() const
@@ -36,31 +36,32 @@ void Texture2D::GenerateMipmap() const
   glGenerateTextureMipmap(id);
 }
 
-void Texture2D::CreateStorage(Texture2DInternalFormat internalFormat, 
-                              i32 width, 
+void Texture2D::CreateStorage(Texture2DInternalFormat internalFormat,
+                              i32 width,
                               i32 height) const
 {
-  u32 mipmapLevels = 1 + std::floor(std::log2(std::max(width, height)));
+  u32 mipmapLevels = 1 + static_cast<u32>(std::floor(std::log2(std::max(width, height))));
+
   glTextureStorage2D(
-    id, 
-    mipmapLevels, 
-    static_cast<u32>(internalFormat),
-    width, 
-    height);
+      id,
+      mipmapLevels,
+      static_cast<u32>(internalFormat),
+      width,
+      height);
 }
 
-void Texture2D::CreateStorageMultisampled(Texture2DInternalFormat internalFormat, 
-                                          i32 samples, 
-                                          i32 width, 
+void Texture2D::CreateStorageMultisampled(Texture2DInternalFormat internalFormat,
+                                          i32 samples,
+                                          i32 width,
                                           i32 height) const
 {
   glTextureStorage2DMultisample(
-    id, 
-    samples, 
-    static_cast<u32>(internalFormat),
-    width, 
-    height, 
-    true);
+      id,
+      samples,
+      static_cast<u32>(internalFormat),
+      width,
+      height,
+      true);
 }
 
 void Texture2D::UpdateStorage(i32 level,
@@ -68,35 +69,34 @@ void Texture2D::UpdateStorage(i32 level,
                               i32 height,
                               Texture2DFormat format,
                               Texture2DSubImageType type,
-                              const void* pixels,
+                              const void *pixels,
                               i32 xoffset,
                               i32 yoffset) const
 {
-  glTextureSubImage2D(id, 
-    level, 
-    xoffset, 
-    yoffset, 
-    width,
-    height,
-    static_cast<u32>(format),
-    static_cast<u32>(type),
-    pixels);
+  glTextureSubImage2D(id,
+                      level,
+                      xoffset,
+                      yoffset,
+                      width,
+                      height,
+                      static_cast<u32>(format),
+                      static_cast<u32>(type),
+                      pixels);
 }
 
 void Texture2D::ClearStorage(i32 level,
                              Texture2DFormat format,
                              Texture2DClearImageType type,
-                             const void* data) const
+                             const void *data) const
 {
-  glClearTexImage(id, 
-    level, 
-    static_cast<u32>(format),
-    static_cast<u32>(type),
-    data);
+  glClearTexImage(id,
+                  level,
+                  static_cast<u32>(format),
+                  static_cast<u32>(type),
+                  data);
 }
 
-
-void Texture2D::LoadImageData(const fs::path& absolute)
+void Texture2D::LoadImageData(const fs::path &absolute)
 {
   if (!fs::exists(absolute))
   {
@@ -105,9 +105,9 @@ void Texture2D::LoadImageData(const fs::path& absolute)
   }
 
   // When to Apply Gamma Correction to Textures?
-  // Gamma correction is only applied to colour diffuse(albedo) textures because the colours 
+  // Gamma correction is only applied to colour diffuse(albedo) textures because the colours
   // must be interpreted in gamma space(sRGB).
-  // Other textures, such as normal, specular, metallic, roughness, are linear data and should not 
+  // Other textures, such as normal, specular, metallic, roughness, are linear data and should not
   // undergo gamma correction.
   bool gammaCorrection = false;
   String filename = absolute.filename().string();
@@ -121,7 +121,7 @@ void Texture2D::LoadImageData(const fs::path& absolute)
   i32 width, height, nChannels;
   Texture2DFormat format = Texture2DFormat::RGB;
   Texture2DInternalFormat internalFormat = Texture2DInternalFormat::RGB8;
-  u8* data = ImageLoader::LoadImageData(absolute, width, height, nChannels);
+  u8 *data = ImageLoader::LoadImageData(absolute, width, height, nChannels);
   if (data)
   {
     // From https://registry.khronos.org/OpenGL-Refpages/gl4/html/glTexStorage2D.xhtml
@@ -130,7 +130,7 @@ void Texture2D::LoadImageData(const fs::path& absolute)
     // GL_SRGB:        gamma correction: yes;  alpha component: no
     // GL_SRGB_ALPHA:  gamma correction: yes;  alpha component: yes
 
-    switch (nChannels) 
+    switch (nChannels)
     {
     case 1:
       internalFormat = Texture2DInternalFormat::R8;
@@ -141,7 +141,7 @@ void Texture2D::LoadImageData(const fs::path& absolute)
       format = Texture2DFormat::RG;
       break;
     case 3:
-      internalFormat = (gammaCorrection ? Texture2DInternalFormat::SRGB8: Texture2DInternalFormat::RGB8);
+      internalFormat = (gammaCorrection ? Texture2DInternalFormat::SRGB8 : Texture2DInternalFormat::RGB8);
       format = Texture2DFormat::RGB;
       break;
     case 4:
@@ -168,43 +168,42 @@ void Texture2D::LoadImageData(const fs::path& absolute)
 void Texture2D::GetTextureImage(i32 level,
                                 Texture2DFormat format,
                                 Texture2DGetImageType type,
-                                i32 buffSize, 
-                                void* pixels) const
+                                i32 buffSize,
+                                void *pixels) const
 {
-  glGetTextureImage(id, 
-    level, 
-    static_cast<u32>(format),
-    static_cast<u32>(type),
-    buffSize, 
-    pixels);
+  glGetTextureImage(id,
+                    level,
+                    static_cast<u32>(format),
+                    static_cast<u32>(type),
+                    buffSize,
+                    pixels);
 }
 
-void Texture2D::SetParameteri(TextureParameteriName name, 
+void Texture2D::SetParameteri(TextureParameteriName name,
                               TextureParameteriParam value) const
 {
   glTextureParameteri(
-    id, 
-    static_cast<u32>(name),
-    static_cast<i32>(value));
+      id,
+      static_cast<u32>(name),
+      static_cast<i32>(value));
 }
 
-void Texture2D::SetParameterfv(TextureParameteriName name, 
-                               f32* values) const
+void Texture2D::SetParameterfv(TextureParameteriName name,
+                               f32 *values) const
 {
   glTextureParameterfv(
-    id,
-    static_cast<u32>(name),
-    values);
+      id,
+      static_cast<u32>(name),
+      values);
 }
 
 void Texture2D::SetCompareFunc(CompareFunc func) const
 {
   glTextureParameteri(
-    id, 
-    static_cast<u32>(TextureParameteriName::COMPARE_FUNC), 
-    static_cast<i32>(func));
+      id,
+      static_cast<u32>(TextureParameteriName::COMPARE_FUNC),
+      static_cast<i32>(func));
 }
-
 
 i32 Texture2D::GetWidth() const
 {
@@ -228,109 +227,109 @@ Texture2DFormat Texture2D::GetFormat(Texture2DInternalFormat internalFormat) con
 {
   switch (internalFormat)
   {
-    case Texture2DInternalFormat::R8:
-    case Texture2DInternalFormat::R8_SNORM:
-    case Texture2DInternalFormat::R16:
-    case Texture2DInternalFormat::R16_SNORM:
-    case Texture2DInternalFormat::R16F:
-    case Texture2DInternalFormat::R32F:
-      return Texture2DFormat::RED;
+  case Texture2DInternalFormat::R8:
+  case Texture2DInternalFormat::R8_SNORM:
+  case Texture2DInternalFormat::R16:
+  case Texture2DInternalFormat::R16_SNORM:
+  case Texture2DInternalFormat::R16F:
+  case Texture2DInternalFormat::R32F:
+    return Texture2DFormat::RED;
 
-    case Texture2DInternalFormat::RG8:
-    case Texture2DInternalFormat::RG8_SNORM:
-    case Texture2DInternalFormat::RG16:
-    case Texture2DInternalFormat::RG16_SNORM:
-    case Texture2DInternalFormat::RG16F:
-    case Texture2DInternalFormat::RG32F:
-      return Texture2DFormat::RG;
+  case Texture2DInternalFormat::RG8:
+  case Texture2DInternalFormat::RG8_SNORM:
+  case Texture2DInternalFormat::RG16:
+  case Texture2DInternalFormat::RG16_SNORM:
+  case Texture2DInternalFormat::RG16F:
+  case Texture2DInternalFormat::RG32F:
+    return Texture2DFormat::RG;
 
-    case Texture2DInternalFormat::RGB8:
-    case Texture2DInternalFormat::RGB8_SNORM:
-    case Texture2DInternalFormat::RGB16:
-    case Texture2DInternalFormat::RGB16_SNORM:
-    case Texture2DInternalFormat::RGB16F:
-    case Texture2DInternalFormat::RGB32F:
-    case Texture2DInternalFormat::R11F_G11F_B10F:
-    case Texture2DInternalFormat::RGB9_E5:
-    case Texture2DInternalFormat::SRGB8:
-      return Texture2DFormat::RGB;
+  case Texture2DInternalFormat::RGB8:
+  case Texture2DInternalFormat::RGB8_SNORM:
+  case Texture2DInternalFormat::RGB16:
+  case Texture2DInternalFormat::RGB16_SNORM:
+  case Texture2DInternalFormat::RGB16F:
+  case Texture2DInternalFormat::RGB32F:
+  case Texture2DInternalFormat::R11F_G11F_B10F:
+  case Texture2DInternalFormat::RGB9_E5:
+  case Texture2DInternalFormat::SRGB8:
+    return Texture2DFormat::RGB;
 
-    case Texture2DInternalFormat::RGBA8:
-    case Texture2DInternalFormat::RGBA8_SNORM:
-    case Texture2DInternalFormat::RGBA16:
-    case Texture2DInternalFormat::RGBA16_SNORM:
-    case Texture2DInternalFormat::RGBA16F:
-    case Texture2DInternalFormat::RGBA32F:
-    case Texture2DInternalFormat::SRGB8_ALPHA8:
-      return Texture2DFormat::RGBA;
+  case Texture2DInternalFormat::RGBA8:
+  case Texture2DInternalFormat::RGBA8_SNORM:
+  case Texture2DInternalFormat::RGBA16:
+  case Texture2DInternalFormat::RGBA16_SNORM:
+  case Texture2DInternalFormat::RGBA16F:
+  case Texture2DInternalFormat::RGBA32F:
+  case Texture2DInternalFormat::SRGB8_ALPHA8:
+    return Texture2DFormat::RGBA;
 
-    case Texture2DInternalFormat::DEPTH_COMPONENT16:
-    case Texture2DInternalFormat::DEPTH_COMPONENT24:
-    case Texture2DInternalFormat::DEPTH_COMPONENT32F:
-      return Texture2DFormat::DEPTH_COMPONENT;
+  case Texture2DInternalFormat::DEPTH_COMPONENT16:
+  case Texture2DInternalFormat::DEPTH_COMPONENT24:
+  case Texture2DInternalFormat::DEPTH_COMPONENT32F:
+    return Texture2DFormat::DEPTH_COMPONENT;
 
-    case Texture2DInternalFormat::DEPTH24_STENCIL8:
-    case Texture2DInternalFormat::DEPTH32F_STENCIL8:
-      return Texture2DFormat::DEPTH_STENCIL;
+  case Texture2DInternalFormat::DEPTH24_STENCIL8:
+  case Texture2DInternalFormat::DEPTH32F_STENCIL8:
+    return Texture2DFormat::DEPTH_STENCIL;
 
-    default:
-      CONSOLE_WARN("Unknown internal format: {:#x} - Defaulting to Texture2DFormat::RED", 
-                   static_cast<u32>(internalFormat));
-      return Texture2DFormat::RED; // Valore di fallback
+  default:
+    CONSOLE_WARN("Unknown internal format: {:#x} - Defaulting to Texture2DFormat::RED",
+                 static_cast<u32>(internalFormat));
+    return Texture2DFormat::RED; // Valore di fallback
   }
 }
 i32 Texture2D::GetNumChannels(Texture2DInternalFormat internalFormat) const
 {
   switch (internalFormat)
   {
-    case Texture2DInternalFormat::R8:
-    case Texture2DInternalFormat::R8_SNORM:
-    case Texture2DInternalFormat::R16:
-    case Texture2DInternalFormat::R16_SNORM:
-    case Texture2DInternalFormat::R16F:
-    case Texture2DInternalFormat::R32F:
-      return 1;
+  case Texture2DInternalFormat::R8:
+  case Texture2DInternalFormat::R8_SNORM:
+  case Texture2DInternalFormat::R16:
+  case Texture2DInternalFormat::R16_SNORM:
+  case Texture2DInternalFormat::R16F:
+  case Texture2DInternalFormat::R32F:
+    return 1;
 
-    case Texture2DInternalFormat::RG8:
-    case Texture2DInternalFormat::RG8_SNORM:
-    case Texture2DInternalFormat::RG16:
-    case Texture2DInternalFormat::RG16_SNORM:
-    case Texture2DInternalFormat::RG16F:
-    case Texture2DInternalFormat::RG32F:
-      return 2;
+  case Texture2DInternalFormat::RG8:
+  case Texture2DInternalFormat::RG8_SNORM:
+  case Texture2DInternalFormat::RG16:
+  case Texture2DInternalFormat::RG16_SNORM:
+  case Texture2DInternalFormat::RG16F:
+  case Texture2DInternalFormat::RG32F:
+    return 2;
 
-    case Texture2DInternalFormat::RGB8:
-    case Texture2DInternalFormat::RGB8_SNORM:
-    case Texture2DInternalFormat::RGB16:
-    case Texture2DInternalFormat::RGB16_SNORM:
-    case Texture2DInternalFormat::RGB16F:
-    case Texture2DInternalFormat::RGB32F:
-    case Texture2DInternalFormat::R11F_G11F_B10F:
-    case Texture2DInternalFormat::RGB9_E5:
-    case Texture2DInternalFormat::SRGB8:
-      return 3;
+  case Texture2DInternalFormat::RGB8:
+  case Texture2DInternalFormat::RGB8_SNORM:
+  case Texture2DInternalFormat::RGB16:
+  case Texture2DInternalFormat::RGB16_SNORM:
+  case Texture2DInternalFormat::RGB16F:
+  case Texture2DInternalFormat::RGB32F:
+  case Texture2DInternalFormat::R11F_G11F_B10F:
+  case Texture2DInternalFormat::RGB9_E5:
+  case Texture2DInternalFormat::SRGB8:
+    return 3;
 
-    case Texture2DInternalFormat::RGBA8:
-    case Texture2DInternalFormat::RGBA8_SNORM:
-    case Texture2DInternalFormat::RGBA16:
-    case Texture2DInternalFormat::RGBA16_SNORM:
-    case Texture2DInternalFormat::RGBA16F:
-    case Texture2DInternalFormat::RGBA32F:
-    case Texture2DInternalFormat::SRGB8_ALPHA8:
-      return 4;
+  case Texture2DInternalFormat::RGBA8:
+  case Texture2DInternalFormat::RGBA8_SNORM:
+  case Texture2DInternalFormat::RGBA16:
+  case Texture2DInternalFormat::RGBA16_SNORM:
+  case Texture2DInternalFormat::RGBA16F:
+  case Texture2DInternalFormat::RGBA32F:
+  case Texture2DInternalFormat::SRGB8_ALPHA8:
+    return 4;
 
-    case Texture2DInternalFormat::DEPTH_COMPONENT16:
-    case Texture2DInternalFormat::DEPTH_COMPONENT24:
-    case Texture2DInternalFormat::DEPTH_COMPONENT32F:
-      return 1;
+  case Texture2DInternalFormat::DEPTH_COMPONENT16:
+  case Texture2DInternalFormat::DEPTH_COMPONENT24:
+  case Texture2DInternalFormat::DEPTH_COMPONENT32F:
+    return 1;
 
-    case Texture2DInternalFormat::DEPTH24_STENCIL8:
-    case Texture2DInternalFormat::DEPTH32F_STENCIL8:
-      return 2;
+  case Texture2DInternalFormat::DEPTH24_STENCIL8:
+  case Texture2DInternalFormat::DEPTH32F_STENCIL8:
+    return 2;
 
-    default:
-      CONSOLE_WARN("Unknown internal format: {:#x} - Defaulting to 1 channel", 
-                   static_cast<u32>(internalFormat));
-      return 1;
+  default:
+    CONSOLE_WARN("Unknown internal format: {:#x} - Defaulting to 1 channel",
+                 static_cast<u32>(internalFormat));
+    return 1;
   }
 }
